Update readme.md
Schema
The data directory contains files named table name.csv where table name is the name used in a CREATE TABLE statement. The result of CREATE TABLE statements is not to create a new file, but simply to synchronize the given TCPH sql schema to an existing .csv file. Delimiters used in these files, (a) vertical-pipe ("|") as a field delimiter, and (b) newlines ("\n") as record delimiters.